feat: add annual insights
